Change srid from uint32 to int32 and update gserialized and TYPMOD handlers to deal with signed SRIDs
git-svn-id: http://svn.osgeo.org/postgis/trunk@6480 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/lwgeom.c
@@ -1630,13 +1630,13 @@ LWGEOM* lwgeom_flip_coordinates(LWGEOM *in)
 	return NULL;
 }
 
-void lwgeom_set_srid(LWGEOM *geom, int srid)
+void lwgeom_set_srid(LWGEOM *geom, int32 srid)
 {
 	int i;
 
 	LWDEBUGF(4,"entered with srid=%d",srid);
 
-	geom->srid = (uint32)srid;
+	geom->srid = srid;
 
 	if ( lwgeom_is_collection(geom) )
 	{
